projects/shop/shop/requirements.txt
asgiref==3.8.1
django==5.2.2
djangorestframework==3.16.0
pillow==11.2.1
psycopg==3.2.9
psycopg2-binary==2.9.10
sqlparse==0.5.3
